LayoutTests/java/lc3/ConvertBoolean/boolean-010-n-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion Preferred argument conversion:  boolean

PASS TEST_CLASS["ambiguous(int)"](true) threw exception TypeError: Result of expression 'TEST_CLASS["ambiguous(int)"]' [undefined] is not a function..
PASS successfullyParsed is true

TEST COMPLETE
 
